✨☀️ Add astrotime sensors based on lat/long
When starting up sensorer you can now specify a `location.lat` and
`location.long` which will be used to calculate sunrise, sunset and
whether it is currently daylight. These points in time are calculated
based on NOAA's astrological time algorithms.

An additional `daylight` timeseries is computed that checks whether
the current time is between sunrise and sunset.